Split MPI resource partitioning routines from controller to mpi_resources.
